requirements.txt
Django
pandas
requests
djangorestframework